Added: Automated GitHub Releases deployment from CI
This set of additions on Travis CI config file, is about deploying
project file on github release services.
When a new Tag is created (ex v2.4.0), it automatically trigs Travis-CI
server to build the last known commit on the selected branch. If the
build was successful, then it uploads the compiled WAR file to
GitHub-Releases-Service, under the specified tag name.